+++ source/others/extF80M_to_ui64.cpp
@@ -74,7 +74,7 @@ extF80M_to_ui64(
     extSig[indexWord(3, 1)] = (uint32_t)sig;
     extSig[indexWord(3, 0)] = 0;
     if (shiftDist) {
-        softfloat_shiftRightJam96M(extSig, shiftDist, extSig);
+        softfloat_shiftRightJam96M(extSig, static_cast<uint8_t>(shiftDist), extSig);
     }
     return softfloat_roundPackMToUI64(sign, extSig, roundingMode, exact);
 }
